feat: add BaseInputField component and enable stand-alone use of Label
affects: @fremtind/jkl-text-input-react, @fremtind/jkl-typography-react
