+++ packages/text-input-react/src/index.ts
@@ -2,3 +2,4 @@ export { TextField } from "./TextField";
 export { TextArea } from "./TextArea";
 export { ActionTextField } from "./ActionTextField";
 export { InlineTextField } from "./InlineTextField";
+export { BaseInputField } from "./BaseInputField";
